Disable app signature check
This will disable the signature check for system applications as well as installable app package. This is important because we are going to mod different system applications heavily and it will be very difficult to keep the original signature intact. If it is not done then the apps modded and signed with different keys will not appear in system as well as app drawer. This is also important so some xperia apps like image enhacement to work properly and to be recognized by the system.